fix: refactor code and send max amount information from context
