+++ react/QuantitySelector.tsx
@@ -1,5 +1,5 @@
 import type { VFC } from 'react'
-import React, { useEffect, useRef, useState } from 'react'
+import React from 'react'
 import classnames from 'classnames'
 import { Loading } from 'vtex.render-runtime'
 import { useCssHandles, applyModifiers } from 'vtex.css-handles'
@@ -10,7 +10,6 @@ import { useItemContext } from './ItemContext'
 import { AVAILABLE, CANNOT_BE_DELIVERED } from './constants/Availability'
 import { opaque } from './utils/opaque'
 import styles from './styles.css'
-import { OrderForm } from 'vtex.order-manager'
 
 const MAX_ITEM_QUANTITY = 99999
 const CSS_HANDLES = ['quantitySelectorContainer'] as const
@@ -29,71 +28,25 @@ function shouldDisableSelector(availability: string | null | undefined) {
   return !enabledSelectorAvailability.includes(availability ?? '')
 }
 
-const { useOrderForm } = OrderForm
-
-const VIRTUAL_MAX_VALUE = 36
-const DEFAULT_GIFT_TABLE_ID = 'default-gift-table-id'
-
 const QuantitySelector: VFC<Props> = ({
   mode = 'default',
   quantitySelectorStep = 'unitMultiplier',
 }) => {
   const { item, loading, onQuantityChange } = useItemContext()
   const handles = useCssHandles(CSS_HANDLES)
-  const orderForm = useOrderForm()
   
-  const mdr = orderForm?.orderForm?.customData?.customApps?.find((app: {id: string}) =>  app?.id === 'mdr')
-  const numart = item.refId
-
-  const [isVirtual, setIsVirtual] = useState(false) 
-  const [physicalInventory, setPhysicalInventory] = useState(MAX_ITEM_QUANTITY) 
-  const isTheFirstTime = useRef(true)
-  const [maxValue, setMaxValue] = useState(MAX_ITEM_QUANTITY) 
-  const [availability, setAvailability] = useState(AVAILABLE) 
-
   if (loading) {
     return <Loading />
   }
 
-  useEffect(() => {
-    const giftTableId = mdr?.fields?.giftTableId ?? DEFAULT_GIFT_TABLE_ID
-
-    if(!numart || !isTheFirstTime.current) return
-
-    const getIsVirtual = async () => {
-      const res = await fetch(`/chapur/v1/sku-virtual/${giftTableId}/${numart}/${item.id}`)
-      const data = await res.json()
-      const virtual = data.virtual ?? false
-      const physicalInventory = data.physicalInventory ?? 0
-
-      setIsVirtual(virtual)
-      setPhysicalInventory(data.physicalInventory)
-      isTheFirstTime.current = false
-    }
-
-    getIsVirtual()
-  }, [numart, mdr, item.id])
-
-  useEffect(() => {
-   setMaxValue(isVirtual && !!mdr ? VIRTUAL_MAX_VALUE : physicalInventory)
-   setAvailability(isVirtual ? item.quantity === maxValue ? CANNOT_BE_DELIVERED : AVAILABLE : physicalInventory === 0 ? CANNOT_BE_DELIVERED  : AVAILABLE)
-  }, [physicalInventory, isVirtual, mdr])
-
   const unitMultiplier =
     quantitySelectorStep === 'singleUnit' ? 1 : item.unitMultiplier ?? undefined
-  
-  const totalUnits = orderForm.orderForm.items
-                      .filter((currentItem: { id: string }) => currentItem.id === item.id)
-                      .reduce((accumulator: number, currentValue: { quantity: number }) => { 
-                        return accumulator + currentValue.quantity
-                      }, 0)
-
 
   if (mode === 'stepper') {
     return (
-      <div
+      item.availability === AVAILABLE ? <div
         className={classnames(
-          opaque(availability),
+          opaque(item.availability),
           handles.quantitySelectorContainer,
           styles.quantity,
           styles.quantityStepper,
@@ -106,21 +59,20 @@ const QuantitySelector: VFC<Props> = ({
         <QuantityStepper
           id={item.id}
           value={item.quantity}
-          maxValue={maxValue}
+          maxValue={item.maxValue ?? MAX_ITEM_QUANTITY}
           onChange={onQuantityChange}
           unitMultiplier={unitMultiplier}
-          disabled={shouldDisableSelector(availability)}
+          disabled={shouldDisableSelector(item.availability)}
           measurementUnit={item.measurementUnit ?? undefined}
-          disabledAddUnitBtn={totalUnits >= maxValue}
         />
-      </div>
+      </div> : null
     )
   }
 
   return (
     <div
       className={classnames(
-        opaque(availability),
+        opaque(item.availability),
         handles.quantitySelectorContainer,
         styles.quantity,
         styles.quantitySelector,
@@ -133,9 +85,9 @@ const QuantitySelector: VFC<Props> = ({
       <Selector
         id={item.id}
         value={item.quantity}
-        maxValue={maxValue}
+        maxValue={item.maxValue ?? MAX_ITEM_QUANTITY}
         onChange={onQuantityChange}
-        disabled={shouldDisableSelector(availability)}
+        disabled={shouldDisableSelector(item.availability)}
         unitMultiplier={unitMultiplier}
         measurementUnit={item.measurementUnit ?? undefined}
       />
